fix: apply code-review findings (coex, websocket, power-off, sensors)
- coex: riabilita lo scan BLE su IP_EVENT_STA_GOT_IP — prima, dopo il
  primo drop WiFi, lo scan restava disabilitato per sempre e i sensori
  power/CSC/HR non venivano piu ritrovati
- ws: client WebSocket creato una sola volta via _coach_ws_start()
  idempotente — coach_init lo ricreava sopra quello del GOT_IP handler
  (leak + doppia connessione)
- coach OFF: esp_wifi_stop() reale, stop del reconnect timer e scan BLE
  riattivato; il disconnect handler non riarma la riconnessione in OFF
- power: esp_deep_sleep(0) risvegliava subito il chip (wakeup timer a
  0 us); ora power_off() rilascia PWR_HOLD con deep sleep come fallback
- sensors: staleness 5s anche per power/HR/cadenza (prima solo speed:
  dopo la disconnessione del power meter il CdA usava potenza congelata)
- sdp810: clamp a [-500,+500] Pa — azzerare i negativi falsava la
  calibrazione dello zero del Pitot
- ble_sensors: distance_m uint16 -> uint32 (niente wrap a 65 km)
- coach start cmd: avvia anche i timer sessione/lap e la schermata TIMER
- qmi8658: rimosso blocco di init duplicato irraggiungibile dopo return
- housekeeping: stack 2048 -> 3072 (ESP_LOGW con float formatting)
- log: PITOT a 10 Hz declassato a LOGD, esp_err_to_name per SDP810,
  rimossa callback lap morta (g_ble_lap_cb mai invocata)

+++ components/ble_central/ble_sensors.c
@@ -217,8 +217,7 @@ static void parse_csc_meas(const uint8_t *d, uint16_t len) {
                 float speed_ms = (float)dr * s_wheel_circ_m * 1024.0f / (float)dt;
                 if (speed_ms > 0.5f && speed_ms < 30.0f) {  /* 1.8 km/h – 108 km/h */
                     s_data.speed_cms  = (uint16_t)(speed_ms * 100.0f);
-                    /* distanza cumulativa — wrappa a ~65535m = 65km */
-                    s_data.distance_m = (uint16_t)((float)wr * s_wheel_circ_m);
+                    s_data.distance_m = (uint32_t)((float)wr * s_wheel_circ_m);
                     s_data.status    |= BLE_SENS_SPEED;
                 }
             }
